+++ fastly/block_fastly_service_logging_newrelicotlp_test.go
@@ -89,7 +89,7 @@ func TestAccFastlyServiceVCL_logging_newrelicotlp_basic(t *testing.T) {
 	})
 }
 
-func testAccCheckFastlyServiceVCLNewRelicOTLPAttributes(service *gofastly.ServiceDetail, newrelicotlp []*gofastly.NewRelicOTLP) resource.TestCheckFunc {
+func testAccCheckFastlyServiceVCLNewRelicOTLPAttributes(service *gofastly.ServiceDetail, newrelicotlp []*gofastly.NewRelicOTLP, serviceType ...string) resource.TestCheckFunc {
 	return func(_ *terraform.State) error {
 		conn := testAccProvider.Meta().(*APIClient).conn
 		newrelicotlpList, err := conn.ListNewRelicOTLP(context.TODO(), &gofastly.ListNewRelicOTLPInput{
@@ -117,6 +117,15 @@ func testAccCheckFastlyServiceVCLNewRelicOTLPAttributes(service *gofastly.Servic
 					// these ahead of time
 					dl.CreatedAt = nil
 					dl.UpdatedAt = nil
+
+					// Ignore VCL attributes for Compute and set to whatever is returned from the API.
+					if len(serviceType) > 0 && serviceType[0] == ServiceTypeCompute {
+						dl.FormatVersion = d.FormatVersion
+						dl.Format = d.Format
+						dl.ResponseCondition = d.ResponseCondition
+						dl.Placement = d.Placement
+						dl.URL = d.URL
+					}
 
 					if diff := cmp.Diff(d, dl); diff != "" {
 						return fmt.Errorf("bad match NewRelic OTLP logging match: %s", diff)
@@ -193,6 +202,77 @@ resource "fastly_service_vcl" "foo" {
   force_destroy = true
 }
 `, name, domain, format, format)
+}
+
+func TestAccFastlyServiceVCL_logging_newrelicotlp_basic_compute(t *testing.T) {
+	var service gofastly.ServiceDetail
+	name := fmt.Sprintf("tf-test-%s", acctest.RandString(10))
+	domain := fmt.Sprintf("fastly-test.%s.com", name)
+
+	log1 := gofastly.NewRelicOTLP{
+		ServiceVersion:   gofastly.ToPointer(1),
+		Name:             gofastly.ToPointer("newrelicotlp-endpoint"),
+		Token:            gofastly.ToPointer("token"),
+		Region:           gofastly.ToPointer(""),
+		Placement:        gofastly.ToPointer("none"),
+		ProcessingRegion: gofastly.ToPointer("none"),
+	}
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		ProviderFactories: testAccProviders,
+		CheckDestroy:      testAccCheckServiceVCLDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccServiceVCLNewRelicOTLPComputeConfig(name, domain),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckServiceExists("fastly_service_compute.foo", &service),
+					testAccCheckFastlyServiceVCLNewRelicOTLPAttributes(&service, []*gofastly.NewRelicOTLP{&log1}, ServiceTypeCompute),
+					resource.TestCheckResourceAttr("fastly_service_compute.foo", "name", name),
+					resource.TestCheckResourceAttr("fastly_service_compute.foo", "logging_newrelicotlp.#", "1"),
+				),
+			},
+		},
+	})
+}
+
+func testAccServiceVCLNewRelicOTLPComputeConfig(name string, domain string) string {
+	return fmt.Sprintf(`
+data "fastly_package_hash" "example" {
+  filename = "./test_fixtures/package/valid.tar.gz"
+}
+
+resource "fastly_service_compute" "foo" {
+  name = "%s"
+
+  domain {
+    name    = "%s"
+    comment = "tf-newrelicotlp-logging"
+  }
+
+  backend {
+    address = "aws.amazon.com"
+    name    = "amazon docs"
+  }
+
+  logging_newrelicotlp {
+    name   = "newrelicotlp-endpoint"
+    token  = "token"
+	placement = "none"
+	format_version = "2"
+	
+
+  }
+
+  package {
+    filename = "test_fixtures/package/valid.tar.gz"
+    source_code_hash = data.fastly_package_hash.example.hash
+  }
+
+  force_destroy = true
+}`, name, domain)
 }
 
 func TestResourceFastlyFlattenNewRelicOTLP(t *testing.T) {