+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - feat(ngwaf/rules): add support for deception and allow_interactive actions ([#1077](https://github.com/fastly/terraform-provider-fastly/pull/1077))
 - feat(compute/healthcheck): add support for healthchecks for Compute services ([#1079](https://github.com/fastly/terraform-provider-fastly/pull/1079))
 - feat(logging): add support for compression to HTTPS logging endpoint ([#1086](https://github.com/fastly/terraform-provider-fastly/pull/1086))
+- feat(compute/logging_newrelicotlp): add support for New Relic OLTP logging for Compute services ([#1095](https://github.com/fastly/terraform-provider-fastly/pull/1095))
 - feat(logging): add support for Period HTTPS logging endpoint ([#](https://github.com/fastly/terraform-provider-fastly/pull/))
 
 ### BUG FIXES: